Fix deposit callback routing
